refactor: 💡 separate auth button and add skeleton
