+++ package.json
@@ -30,6 +30,7 @@
     "@nextui-org/link": "2.0.32",
     "@nextui-org/listbox": "2.1.21",
     "@nextui-org/navbar": "2.0.33",
+    "@nextui-org/skeleton": "^2.0.31",
     "@nextui-org/snippet": "2.0.38",
     "@nextui-org/spinner": "^2.0.30",
     "@nextui-org/switch": "2.0.31",
